Changed type of attributes from double to int

+++ src/domain/body/fallingBody/Gift.java
@@ -4,10 +4,10 @@ public class Gift extends FallingBody{
     public String giftName;
     public int duration;
 
-    public Gift(double x_coordinates,
-                double y_coordinates,
-                double length,
-                double width,
+    public Gift(int x_coordinates,
+                int y_coordinates,
+                int length,
+                int width,
                 String giftName,
                 int duration) {
         super(x_coordinates, y_coordinates, length, width);
